Separate FD First implementation from common CFD code
