+++ src/core/algorithms/cfd/CMakeLists.txt
@@ -1,27 +1,19 @@
 add_subdirectory(cfd_verifier)
-
-set(NAME cfd.util)
-desbordante_add_lib(NAME OBJECT)
-target_sources(${NAME} PRIVATE util/partition_util.cpp util/prefix_tree.cpp)
-target_link_libraries(${NAME} PRIVATE Boost::headers)
+add_subdirectory(fd_first)
 
 set(NAME cfd.model)
 desbordante_add_lib(NAME OBJECT)
-target_sources(
-    ${NAME} PRIVATE model/cfd_relation_data.cpp model/partition_tidlist.cpp model/raw_cfd.cpp
-)
+target_sources(${NAME} PRIVATE model/cfd_relation_data.cpp model/raw_cfd.cpp)
 target_link_libraries(
-    ${NAME} PRIVATE ${DESBORDANTE_PREFIX}::cfd::util ${DESBORDANTE_PREFIX}::model::table
-                    spdlog::spdlog_header_only Boost::headers
+    ${NAME} PRIVATE ${DESBORDANTE_PREFIX}::model::table spdlog::spdlog_header_only Boost::headers
 )
 
 set(NAME cfd)
 desbordante_add_lib(NAME)
-target_sources(${NAME} PRIVATE cfd_discovery.cpp fd_first_algorithm.cpp)
+target_sources(${NAME} PRIVATE cfd_discovery.cpp)
 target_link_libraries(
     ${NAME}
     PUBLIC ${DESBORDANTE_PREFIX}::config
-    PRIVATE ${DESBORDANTE_PREFIX}::cfd::model ${DESBORDANTE_PREFIX}::cfd::util
-            ${DESBORDANTE_PREFIX}::algos ${DESBORDANTE_PREFIX}::util spdlog::spdlog_header_only
-            Boost::headers
+    PRIVATE ${DESBORDANTE_PREFIX}::cfd::model ${DESBORDANTE_PREFIX}::algos
+            ${DESBORDANTE_PREFIX}::util spdlog::spdlog_header_only Boost::headers
 )
