+++ src/core/algorithms/cfd/cfd_discovery.cpp
@@ -1,12 +1,9 @@
 #include "core/algorithms/cfd/cfd_discovery.h"
 
-#include "core/algorithms/cfd/util/set_util.h"
 #include "core/config/equal_nulls/option.h"
 #include "core/config/names_and_descriptions.h"
 #include "core/config/option_using.h"
 #include "core/config/tabular_data/input_table/option.h"
-
-// see algorithms/cfd/LICENSE
 
 namespace algos::cfd {
