+++ src/core/algorithms/cfd/fd_first/fd_first_algorithm.cpp
@@ -1,4 +1,4 @@
-#include "core/algorithms/cfd/fd_first_algorithm.h"
+#include "core/algorithms/cfd/fd_first/fd_first_algorithm.h"
 
 #include <algorithm>
 #include <iterator>
@@ -7,15 +7,15 @@
 
 #include <boost/unordered_map.hpp>
 
-#include "core/algorithms/cfd/util/partition_util.h"
-#include "core/algorithms/cfd/util/set_util.h"
+#include "core/algorithms/cfd/fd_first/util/partition_util.h"
+#include "core/algorithms/cfd/fd_first/util/set_util.h"
 #include "core/config/equal_nulls/option.h"
 #include "core/config/exceptions.h"
 #include "core/config/names_and_descriptions.h"
 #include "core/config/option_using.h"
 #include "core/util/logger.h"
 
-// see algorithms/cfd/LICENSE
+// see algorithms/cfd/fd_first/LICENSE
 
 namespace algos::cfd {
