+++ src/core/algorithms/cfd/fd_first/fd_first_algorithm.h
@@ -1,12 +1,17 @@
 #pragma once
 
-#include "core/algorithms/cfd/cfd_discovery.h"
-#include "core/algorithms/cfd/enums.h"
-#include "core/algorithms/cfd/miner_node.h"
-#include "core/algorithms/cfd/model/partition_tidlist.h"
-#include "core/algorithms/cfd/util/prefix_tree.h"
+#include <map>
+#include <set>
+#include <unordered_map>
 
-// see algorithms/cfd/LICENSE
+#include "core/algorithms/cfd/cfd_discovery.h"
+#include "core/algorithms/cfd/fd_first/enums.h"
+#include "core/algorithms/cfd/fd_first/miner_node.h"
+#include "core/algorithms/cfd/fd_first/model/partition_tidlist.h"
+#include "core/algorithms/cfd/fd_first/model/types.h"
+#include "core/algorithms/cfd/fd_first/util/prefix_tree.h"
+
+// see algorithms/cfd/fd_first/LICENSE
 
 namespace algos::cfd {
