+++ src/core/algorithms/cfd/fd_first/miner_node.h
@@ -4,7 +4,7 @@
 
 #include "core/algorithms/cfd/model/cfd_types.h"
 
-// see algorithms/cfd/LICENSE
+// see algorithms/cfd/fd_first/LICENSE
 
 namespace algos::cfd {
