+++ src/core/algorithms/cfd/fd_first/util/set_util.h
@@ -5,7 +5,7 @@
 #include <functional>
 #include <vector>
 
-// see algorithms/cfd/LICENSE
+// see algorithms/cfd/fd_first/LICENSE
 
 namespace algos::cfd {
