+++ src/core/config/descriptions.cpp
@@ -2,7 +2,7 @@
 
 #include <algorithm>
 
-#include "core/algorithms/cfd/enums.h"
+#include "core/algorithms/cfd/fd_first/enums.h"
 #include "core/algorithms/cind/types.h"
 #include "core/algorithms/fd/afd_metric/afd_metric.h"
 #include "core/algorithms/fd/tane/enums.h"
