+++ src/python_bindings/CMakeLists.txt
@@ -84,7 +84,7 @@ desbordante_add_bind(
     cfd/bind_cfd.cpp
     cfd/bind_cfd_verification.cpp
     LIBS
-    ${DESBORDANTE_PREFIX}::cfd
+    ${DESBORDANTE_PREFIX}::cfd::fdfirst
     ${DESBORDANTE_PREFIX}::cfd::verifier
     Boost::headers
 )
