+++ src/python_bindings/py_util/py_to_any.cpp
@@ -14,7 +14,7 @@
 #include <pybind11/stl/filesystem.h>
 
 #include "core/algorithms/algebraic_constraints/bin_operation_enum.h"
-#include "core/algorithms/cfd/enums.h"
+#include "core/algorithms/cfd/fd_first/enums.h"
 #include "core/algorithms/cfd/model/raw_cfd.h"
 #include "core/algorithms/cind/types.h"
 #include "core/algorithms/dd/dd.h"
