+++ src/tests/unit/test_fdfirst.cpp
@@ -4,8 +4,8 @@
 #include <gtest/gtest.h>
 
 #include "core/algorithms/algo_factory.h"
-#include "core/algorithms/cfd/enums.h"
-#include "core/algorithms/cfd/fd_first_algorithm.h"
+#include "core/algorithms/cfd/fd_first/enums.h"
+#include "core/algorithms/cfd/fd_first/fd_first_algorithm.h"
 #include "core/config/names.h"
 #include "tests/common/all_csv_configs.h"
